Show physico-chemical metadata in dashboard
