+++ bims/templates/water_temperature_single_site.html
@@ -407,6 +407,7 @@
     <script src="{% static "js/utils/table_download.js" %}"></script>
     <script src="{% static "js/utils/svg_download.js" %}"></script>
     <script src="{% static "js/sass_dashboard_close.js" %}"></script>
+    <script src="{% static "js/metadata_table.js" %}"></script>
     <script src="{% static "js/non_requirejs/water_temperature.js" %}"></script>
 
 {% endblock %}
