+++ sass/templates/sass_dashboard_single_site.html
@@ -477,6 +477,7 @@
     <script src="{% static "js/utils/download_csv.js" %}"></script>
     <script src="{% static "js/utils/table_download.js" %}"></script>
     <script src="{% static "js/utils/svg_download.js" %}"></script>
+    <script src="{% static "js/metadata_table.js" %}"></script>
     <script src="{% static "js/sass_dashboard.js" %}"></script>
     <script src="{% static "js/sass_dashboard_close.js" %}"></script>
